+++ packages/hardhat/contracts/interfaces/ICultFactory.sol
@@ -38,6 +38,6 @@ interface ICultFactory {
         string memory _name,
         string memory _symbol,
         bytes32[] calldata _merkleRoots,
-        uint16 _airdropPercent
+        uint32 _airdropPercent
     ) external payable returns (address);
 }
